% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_clusterplot.R
\name{coef_clusterplot}
\alias{coef_clusterplot}
\title{A function to plot coefficients in Logit model with cluster sd .}
\usage{
coef_clusterplot(ModelResults, varnames = NULL, data, clusterid)
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{varnames}{character vector. A vector contains the name of variables}

\item{data}{data.frame. The name of the data frame in logit model}

\item{clusterid}{character. The cluster id}
}
\description{
This function creates a  coefficient plot for Logit model
}
