% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_clusterplot.R
\name{density_clusterplot}
\alias{density_clusterplot}
\title{A function to plot Predicted Probability in Logit model with cluster sd .}
\usage{
density_clusterplot(ModelResults, n.sim = 1000, varname, data, clusterid,
  label = c("lab1", "lab2"))
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{n.sim}{numeric. Specify the number of simulations}

\item{varname}{character vector. A vector contains the name of variables}

\item{data}{data.frame. The name of the data frame in logit model}

\item{clusterid}{character. The cluster id}

\item{label}{vector. Specify the labels}
}
\description{
This function creates a  density plot for a binary variable in Logit model using simulations
}
