% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ridgeplot.R
\name{density_ridgeplot}
\alias{density_ridgeplot}
\title{A function to plot first difference in Logit model with cluster sd .}
\usage{
density_ridgeplot(ModelResults, n.sim = 1000, varname, data, val1, val2,
  clusterid)
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{n.sim}{numeric. Specify the number of simulations}

\item{varname}{character vector. A vector contains the name of variables}

\item{data}{data.frame. The name of the data frame in logit model}

\item{val1}{numeric. Specify the min value of percentile}

\item{val2}{numeric. Specify the max value of percentile}

\item{clusterid}{character. The cluster id}
}
\description{
This function creates a ridge density plot for variables in Logit model using simulations
}
