% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FDEffect.R
\name{plot_FDEffect}
\alias{plot_FDEffect}
\title{A function to plot the first difference in the predicted probability of the interaction between a binary and interval variables in Logit model with cluster sd .}
\usage{
plot_FDEffect(ModelResults, n.sim = 1000, data, clusterid, varname1, varname2,
  val1, val2, intervals, xlabs, ylabs)
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{n.sim}{numeric. Specify the number of simulations}

\item{data}{data.frame. The name of the data frame in logit model}

\item{clusterid}{character. The cluster id}

\item{varname1}{character. A binary variable}

\item{varname2}{character. An interval vaiable}

\item{val1}{numeric. Specify the min value}

\item{val2}{numeric. Specify the max value}

\item{intervals}{numeric. Specify the intervals for the sequence}

\item{xlabs}{character. The x label for the plot}

\item{ylabs}{character. The y label for the plot}
}
\description{
This function creates a facet plot for the interaction terms in Logit model using simulations
}
