% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{A function to plot ROC plots
This function creates a ggplot for ROC}
\usage{
plot_roc(ModelResults, linetypes = c("solid", "dotted"), interval)
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{linetypes}{character. A vector of charactors indicting the line types}

\item{interval}{numeric. Specify the intervals for the sequence}
}
\description{
A function to plot ROC plots
This function creates a ggplot for ROC
}
