% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cluster.R
\name{sim_cluster}
\alias{sim_cluster}
\title{A function to simulate coefficient in Logit model with cluster sd .}
\usage{
sim_cluster(ModelResults, n.sim = 1000, data, clusterid)
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{n.sim}{numeric. Specify the number of simulations}

\item{data}{data.frame. The name of the data frame in logit model}

\item{clusterid}{character. The cluster id}
}
\description{
This function simulate the post esitmation distribution
}
