% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cluster.R
\name{table_cluster}
\alias{table_cluster}
\title{A function to make a latex table from Logit model with cluster sd .}
\usage{
table_cluster(ModelResults, varDef, digs = 3, robustse = TRUE, data,
  clusterid)
}
\arguments{
\item{ModelResults}{list. Specify model name of the logit model.}

\item{varDef}{character vector. A vector contains the name of variables}

\item{data}{data.frame. The name of the data frame in logit model}

\item{clusterid}{character. The cluster id}
}
\description{
This function creates a  latex table for Logit model
}
